requirements.txt

# Realtime ingestion
praw
google-api-python-client
SQLAlchemy
python-dotenv
requests
transformers
torch
scikit-learn
bertopic
umap-learn
keybert
spacy[en_core_web_sm]
streamlit
wordcloud
matplotlib
pandas
streamlit-autorefresh

# Optional: make sure spacy model is downloaded
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl